Re-organized test package
